More cleanup; Adds TLAPACK_PREFERRED_MATRIX macro
