+++ examples/cpp_visualizer/cpp_visualizer_example.cpp
@@ -7,6 +7,7 @@
 // <T>LAPACK is free software: you can redistribute it and/or modify it under
 // the terms of the BSD 3-Clause license. See the accompanying LICENSE file.
 
+#include <tlapack/plugins/legacyArray.hpp>
 #include <tlapack.hpp>
 #include <tlapack/plugins/debugutils.hpp>
