+++ examples/eigenvalues/example_eigenvalues.cpp
@@ -8,6 +8,7 @@
 // the terms of the BSD 3-Clause license. See the accompanying LICENSE file.
 
 #include <tlapack/plugins/stdvector.hpp>
+#include <tlapack/plugins/legacyArray.hpp>
 #include <tlapack.hpp>
 
 #include <memory>
