+++ examples/potrf/example_potrf.cpp
@@ -12,6 +12,7 @@
 #include <chrono>   // for high_resolution_clock
 
 // <T>LAPACK
+#include <tlapack/plugins/legacyArray.hpp>
 #include <tlapack.hpp>
 
 #ifndef USE_MKL
