+++ test/src/test_optBLAS.cpp
@@ -8,9 +8,8 @@
 // the terms of the BSD 3-Clause license. See the accompanying LICENSE file.
 
 #include <catch2/catch_template_test_macros.hpp>
-#include <testdefinitions.hpp>
 
-#include <complex>
+#include "testutils.hpp"
 
 using namespace tlapack;
